+++ tests/test/dashboard.test.js
@@ -7,10 +7,6 @@ describe('The Unlock Dashboard', () => {
 
   const testLockAddress = '0x5Cd3FC283c42B4d5083dbA4a6bE5ac58fC0f0267'
 
-  function lockSelector(name) {
-    return `#${name}_${testLockAddress}`
-  }
-
   it('should load the creator dashboard', async () => {
     await expect(page).toMatch('Creator Dashboard')
   })
@@ -48,7 +44,7 @@ describe('The Unlock Dashboard', () => {
       await expect(page).toMatch('Code snippet')
     }, 25000)
 
-    xit('has a Preview dashboard button', async () => {
+    it('has a Preview dashboard button', async () => {
       await page.waitForSelector(`#PreviewButton_${testLockAddress}`)
       await expect(page).toMatchElement(`#PreviewButton_${testLockAddress}`)
     }, 25000)
@@ -77,45 +73,5 @@ describe('The Unlock Dashboard', () => {
       await page.waitFor(3500)
       await expect(page).toMatch('Updated Lock Name')
     })
-  })
-  
-  describe('Paywall demo', () => {
-    beforeAll(async () => {
-      await Promise.all([
-        page.goto(url(`/demo/${testLockAddress}`)),
-        page.waitForNavigation(),
-      ])
-    })
-    it('should display a lock on the demo page with a paywall', async () => {
-      await page.waitForFunction(() => window.frames.length)
-      const paywallIframe = page.mainFrame().childFrames()[0]
-      await paywallIframe.waitForSelector(lockSelector('Lock'))
-      await paywallIframe.waitForFunction((ethPriceSelector) => {
-        const eth = document.querySelector(ethPriceSelector)
-        if (!eth) return false
-        return eth.innerText === '0.33 ETH'
-      }, {}, lockSelector('EthPrice'))
-    }, 15000)
-    it('clicking the lock purchases a key', async () => {
-      const paywallIframe = page.mainFrame().childFrames()[0]
-      const paywallBody = await paywallIframe.$('body')
-      await expect(paywallBody).toClick(lockSelector('PurchaseKey'))
-      await paywallIframe.waitForFunction((footerSelector) => {
-        const footer = document.querySelector(footerSelector)
-        if (!footer) return false
-        return footer.innerText === 'Payment Pending'
-      }, {}, 'footer')
-    }, 15000)
-    it('after key purchase, unlocked flag appears', async () => {
-      await page.reload()
-      await page.waitFor(1000)
-      await page.waitForFunction(() => window.frames.length)
-      const paywallIframe = page.mainFrame().childFrames()[0]
-      await paywallIframe.waitForFunction(() => {
-        const unlockFlag = document.querySelector('#UnlockFlag')
-        if (!unlockFlag) return false
-        return true
-      })
-    }, 15000)
   })
 })
